Fix the flowchart of the game
Now the flowchart of the game is:
(1) State Machine -- do some movement
(2) Collision Detection --check collision
(3) Collision Response -- response to collision
The collision response must be done in the same frame with the collision
detection and it must be done after the collision detection. So the new
flowchart is well

We add the OnCollision method to MLState, so every state must implement
this method, and it can be called by the collision response system.